fix: 多线程访问 RequestGroupMan 导致的数据异常

+++ src/storage/WrDiskCache.h
@@ -58,18 +58,18 @@ public:
   // Updates the already added entry |ent|. The |delta| means how many
   // bytes is increased in this update. If the size is reduced, use
   // negative value.
-  bool update(WrDiskCacheEntry* ent, ssize_t delta);
+  bool update(WrDiskCacheEntry* ent, int64_t delta);
   // Evicts entries from storage so that total size of cache is kept
   // under the limit.
   void ensureLimit();
-  size_t getSize() const { return total_; }
+  int64_t getSize() const { return total_; }
 
 private:
   typedef std::set<WrDiskCacheEntry*, DerefLess<WrDiskCacheEntry*>> EntrySet;
   // Maximum number of bytes the storage can cache.
-  size_t limit_;
+  int64_t limit_;
   // Current number of bytes cached.
-  size_t total_;
+  int64_t total_;
   EntrySet set_;
   int64_t clock_;
 };
